refactor: move `ParamType` type inside `fuels-types`
