+++ packages/fuels-types/src/lib.rs
@@ -7,6 +7,7 @@ use serde::{Deserialize, Serialize};
 use strum_macros::ToString;
 
 pub mod errors;
+pub mod param_types;
 /// Fuel ABI representation in JSON, originally specified here:
 ///
 /// https://github.com/FuelLabs/fuel-specs/blob/master/specs/protocol/abi.md
