+++ tools/fuels-abi-cli/src/main.rs
@@ -1,8 +1,7 @@
 use fuels_core::code_gen::abigen::Abigen;
 use fuels_core::json_abi::parse_param;
 use fuels_core::json_abi::ABIParser;
-use fuels_core::ParamType;
-use fuels_types::Property;
+use fuels_types::{param_types::ParamType, Property};
 
 use std::fs;
 use std::path::PathBuf;
